Fix sonar bugs - precision and NPE

+++ dmn-core/src/main/java/com/gs/dmn/feel/lib/UniformJavaTimeFEELLib.java
@@ -764,7 +764,7 @@ public class UniformJavaTimeFEELLib extends BaseFEELLib<BigDecimal, ZonedDateTim
     public Duration timeOffset(ZonedDateTime date) {
         // timezone offset in seconds
         int secondsOffset = date.getOffset().getTotalSeconds();
-        return DATA_TYPE_FACTORY.newDuration(secondsOffset * 1000);
+        return DATA_TYPE_FACTORY.newDuration((long) secondsOffset * 1000);
     }
 
     public String timezone(ZonedDateTime date) {
